+++ src/components/manage/Tiles/Description/Edit.jsx
@@ -134,7 +134,10 @@ export default class Edit extends Component {
           onChange={this.onChange}
           editorState={this.state.editorState}
           blockRenderMap={extendedBlockRenderMap}
-          handleReturn={() => {
+          handleReturn={e => {
+            if (e.shiftKey) {
+              return 'not-handled';
+            }
             this.props.onSelectTile(
               this.props.onAddTile('text', this.props.index + 1),
             );